refactor(trigger): move secret_provider + template_preflight yaml to testdata/
Follow-up to the previous fixtures refactor: pulls the last two e2e
tests off Go-embedded YAML and onto checked-in fixtures.

Both tests needed dynamic YAML values outside expandSpec's allowlist
(permissions.net, trigger.before[].overrides), so straight LoadDir
wasn't enough. Adds loadFixtureTpl: a sibling of loadFixture that
reads task.yaml, substitutes {{KEY}} placeholders from a vars map,
copies the rendered yaml + task.ts into a t.TempDir(), then calls
LoadDirWithVars with the same vars (so the loader's ${KEY} expansion
sees them too). The two substitution layers use distinct marker
syntaxes so neither clobbers the other — templates with literal
${VAR} content for runtime expansion survive the fixture-load pass
intact.

Drops the writeRendererTask / writeVerifierTask / writeDaemonTask /
quoteYAML helpers and the renderTaskScript / verifyTaskScript const
strings, all of which existed only to manufacture YAML in Go.

Net -189 lines in pkg/trigger/, +6 fixture files. All five e2e tests
still pass.
